extracts to another file
