+++ fastlane/github_api/api.rb
@@ -66,61 +66,3 @@ def post_new_comment(comment_url, body, github_token)
         Fastlane::UI.user_error!("Error posting new comment.\nCode: #{response.code}\nBody: #{response.read_body}")
     end
 end
-
-def get_pull_request_id(pull_request_number, github_token)
-    url = URI('https://api.github.com/graphql')
-    http = Net::HTTP.new(url.host, url.port)
-    http.use_ssl = true
-
-    request = Net::HTTP::Post.new(url.request_uri)
-    request.basic_auth(github_token, '')
-
-    query_id = <<-GRAPHQL
-    query QueryId {
-        repository(name: "revenuecat-docs", owner: "RevenueCat") {
-            pullRequest(number: #{pull_request_number}) {
-                id
-            }
-        }
-    }
-    GRAPHQL
-
-    request.body = { query: query_id }.to_json
-    response = http.request(request)
-    if response.is_a?(Net::HTTPSuccess)
-        pull_request_id = JSON.parse(response.body)["data"]["repository"]["pullRequest"]["id"]
-        if pull_request_id.nil?
-            Fastlane::UI.user_error!("Error querying pull request. Pull request ID is nil.")
-        else
-            Fastlane::UI.message("Pull request ID queried successfully!")
-            return pull_request_id
-        end
-    else
-        Fastlane::UI.user_error!("Error querying pull request.\nCode: #{response.code}\nBody: #{response.read_body}")
-    end
-end
-
-def enable_auto_merge(pull_request_id, github_token)
-    url = URI('https://api.github.com/graphql')
-    http = Net::HTTP.new(url.host, url.port)
-    http.use_ssl = true
-
-    request = Net::HTTP::Post.new(url.request_uri)
-    request.basic_auth(github_token, '')
-
-    query_id = <<-GRAPHQL
-    mutation EnableAutoMerge {
-        enablePullRequestAutoMerge(input: {pullRequestId: "#{pull_request_id}", mergeMethod: SQUASH}) {
-            clientMutationId
-        }
-    }
-    GRAPHQL
-
-    request.body = { query: query_id }.to_json
-    response = http.request(request)
-    if response.is_a?(Net::HTTPSuccess)
-        Fastlane::UI.message("Auto merge enabled successfully!")
-    else
-        Fastlane::UI.user_error!("Failed to enable auto merge.\nCode: #{response.code}\nBody: #{response.read_body}")
-    end
-end
